LAB 5

+++ lab3/lab3server/src/main/java/com/aioisisi/lab2/security/TokenAuthenticationFilter.java
@@ -32,9 +32,7 @@ public class TokenAuthenticationFilter extends OncePerRequestFilter {
             String jwt = getJwtFromRequest(request);
 
             if (StringUtils.hasText(jwt) && tokenProvider.validateToken(jwt)) {
-
-                //todo (mistake may be here)
-                Integer userId = Math.toIntExact(tokenProvider.getUserIdFromToken(jwt));
+                Integer userId = tokenProvider.getUserIdFromToken(jwt);
 
                 UserDetails userDetails = customUserDetailsService.loadUserById(userId);
                 UsernamePasswordAuthenticationToken authentication = new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
